WafflePedia/requirements.txt
asgiref==3.7.2
astroid==3.0.2
black==23.12.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
defusedxml==0.7.1
dill==0.3.7
2
Django==4.2.8
django-allauth==0.59.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pycparser==2.21
PyJWT==2.8.0
pylint==3.0.3
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
urllib3==2.1.0
